+++ project/flood-fill/headers/water_surface_manager.hpp
@@ -24,7 +24,7 @@ private:
   WaterSurfacePtr current;
   glm::vec3 nextPos;
   int nextColorMask;
-  float hasNextTimer;
+  double hasNextTimer;
 };
 
 #endif
