+++ project/flood-fill/source/boss_level.cpp
@@ -128,7 +128,7 @@ void BossLevel::update(){
     if (!Menu::isActive()) {
         glm::vec4 titleColor = levelTitle->getColor();
         if(titleColor.w > 0){
-            titleColor.w -= TimeManager::getDeltaTime()*0.3;
+            titleColor.w -= (float)TimeManager::getDeltaTime() * 0.3f;
             levelTitle->setColor(titleColor);
         }
     }
